requirements.txt
httpx>=0.28.0
rich>=13.9.0
fpdf2>=2.8.0
pydantic>=2.10.0
thefuzz[speedup]>=0.22.0
tqdm>=4.67.0
inquirer>=3.4.0
pillow>=11.0.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-respx>=0.21.1
fastapi>=0.115.0
uvicorn>=0.34.0
easyocr>=1.7.1
argostranslate>=1.9.1
torch>=2.2.0
